+++ app/src/main/java/to/bitkit/env/Env.kt
@@ -246,7 +246,7 @@ internal object Env {
 @Suppress("ConstPropertyName")
 object Defaults {
     /** Default Bolt11 invoice expiry in seconds. */
-    const val bolt11InvoiceExpirySeconds = 3_600u
+    const val bolt11ExpirySec = 86_400u
 
     /** Recommended transaction base fee in sats */
     const val recommendedBaseFee = 256u